+++ component/otelcol/exporter/loadbalancing/loadbalancing.go
@@ -137,8 +137,9 @@ func (otlpConfig OtlpConfig) Convert() otlpexporter.Config {
 
 // ResolverSettings defines the configurations for the backend resolver
 type ResolverSettings struct {
-	Static *StaticResolver `river:"static,block,optional"`
-	DNS    *DNSResolver    `river:"dns,block,optional"`
+	Static     *StaticResolver     `river:"static,block,optional"`
+	DNS        *DNSResolver        `river:"dns,block,optional"`
+	Kubernetes *KubernetesResolver `river:"kubernetes,block,optional"`
 }
 
 func (resolverSettings ResolverSettings) Convert() loadbalancingexporter.ResolverSettings {
@@ -152,6 +153,11 @@ func (resolverSettings ResolverSettings) Convert() loadbalancingexporter.Resolve
 	if resolverSettings.DNS != nil {
 		dnsResolver := resolverSettings.DNS.Convert()
 		res.DNS = &dnsResolver
+	}
+
+	if resolverSettings.Kubernetes != nil {
+		kubernetesResolver := resolverSettings.Kubernetes.Convert()
+		res.K8sSvc = &kubernetesResolver
 	}
 
 	return res
@@ -196,6 +202,29 @@ func (dnsResolver *DNSResolver) Convert() loadbalancingexporter.DNSResolver {
 		Port:     dnsResolver.Port,
 		Interval: dnsResolver.Interval,
 		Timeout:  dnsResolver.Timeout,
+	}
+}
+
+// KubernetesResolver defines the configuration for the k8s resolver
+type KubernetesResolver struct {
+	Service string  `river:"service,attr"`
+	Ports   []int32 `river:"ports,attr,optional"`
+}
+
+var _ river.Defaulter = &KubernetesResolver{}
+
+// SetToDefault implements river.Defaulter.
+func (args *KubernetesResolver) SetToDefault() {
+	if args == nil {
+		args = &KubernetesResolver{}
+	}
+	args.Ports = []int32{4317}
+}
+
+func (k8sSvcResolver *KubernetesResolver) Convert() loadbalancingexporter.K8sSvcResolver {
+	return loadbalancingexporter.K8sSvcResolver{
+		Service: k8sSvcResolver.Service,
+		Ports:   append([]int32{}, k8sSvcResolver.Ports...),
 	}
 }