+++ CHANGELOG.md
@@ -156,6 +156,9 @@ v0.37.4 (2023-11-06)
 - Fix a bug where reloading the configuration of a `loki.write` component lead
   to a panic. (@tpaschalis)
 
+- Added Kubernetes service resolver to static node's loadbalancing exporter 
+  and to Flow's `otelcol.exporter.loadbalancing`. (@ptodev)
+
 v0.37.3 (2023-10-26)
 -----------------